+++ docker/install.sh
@@ -3,9 +3,11 @@
 dotpath=$(dirname $(readlink -f "$0"))
 . "$dotpath/../util.sh"
 
-if [ -e "$dotpath/completion.zsh" ]; then
-  success "docker completion already installed"
-else
-  info "downloading docker completion"
-  curl -sSL https://raw.githubusercontent.com/docker/cli/master/contrib/completion/zsh/_docker -o "$dotpath/completion.zsh"
+if command_exists docker; then
+  if [ -e "$dotpath/completion.zsh" ]; then
+    success "docker completion already installed"
+  else
+    info "downloading docker completion"
+    curl -sSL https://raw.githubusercontent.com/docker/cli/master/contrib/completion/zsh/_docker -o "$dotpath/completion.zsh"
+  fi
 fi
